# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the Vipin LMS platform.",
        "title": "Vipin LMS API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/courses/{id}/videos": {
            "get": {
                "description": "Owner sees every video. Students see ready videos of published lessons. Users previewing a published course they are not enrolled in see them too, with locked=true on the ones they cannot stream.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "List a course's videos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Videos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.VideosResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Course not found, or not visible to this user",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/lessons/{id}/videos": {
            "get": {
                "description": "Owner sees every video. Students see ready videos of published lessons. Users previewing a published course they are not enrolled in see them too, with locked=true on the ones they cannot stream.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "List a lesson's videos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Videos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.VideosResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Lesson not found, or not visible to this user",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Course owner (instructor or admin) gets a signed URL to PUT a video file (mp4, mov, mkv or webm, max 5 GB) straight to storage. Send the returned headers unchanged with the PUT, then call POST /videos/{id}/confirm.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Start a video upload on a lesson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File name and whether the video is free to watch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.requestVideoUploadBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Upload started",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.VideoUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid body or unsupported file name",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Not an instructor",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Lesson not found, or not in a course this user teaches",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "503": {
                        "description": "File storage is disabled (GCS_ENABLE is not true)",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/videos/{id}": {
            "get": {
                "description": "The uploader (who must still own the course) polls this until status is ready or failed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get a video's upload and transcoding status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.VideoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Not an instructor",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/videos/{id}/confirm": {
            "post": {
                "description": "Call after the file was PUT to the signed URL. Checks the file and starts transcoding in the background; poll GET /videos/{id} until status is ready.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Confirm a video upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Transcoding queued",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.VideoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "File not uploaded yet, empty, or wrong type",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Not an instructor",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "409": {
                        "description": "Video is not awaiting upload",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "503": {
                        "description": "Storage disabled, or transcoding queue busy (retry)",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/videos/{id}/hls/{file}": {
            "get": {
                "description": "Public: the signature in the query (from GET /videos/{id}/stream) is the credential. Returns the .m3u8 with every URI rewritten to a signed URL: child playlists point back here, segments go straight to the CDN. Open to any origin.",
                "produces": [
                    "application/vnd.apple.mpegurl"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get a signed HLS playlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Playlist file, e.g. manifest.m3u8",
                        "name": "file",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From stream URL",
                        "name": "URLPrefix",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From stream URL",
                        "name": "Expires",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From stream URL",
                        "name": "KeyName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From stream URL",
                        "name": "Signature",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Playlist",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Invalid or expired signature",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Video or playlist not found",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.VideoResponse": {
            "type": "object",
            "properties": {
                "video": {
                    "$ref": "#/definitions/models.Video"
                }
            }
        },
...
        "handlers.VideoUploadResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "uploadUrl": {
                    "type": "string"
                },
                "video": {
                    "$ref": "#/definitions/models.Video"
                }
            }
        },
        "handlers.VideosResponse": {
            "type": "object",
            "properties": {
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Video"
                    }
                }
            }
        },
        "handlers.requestVideoUploadBody": {
            "type": "object",
            "properties": {
                "fileName": {
                    "type": "string"
                },
                "isFree": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}